+++ CHANGELOG.md
@@ -53,6 +53,7 @@ The next release will require at least [Go 1.25].
 
 - Fix panic when passing nil `TracerProvider` or `MeterProvider` to `WithTracerProvider` or `WithMeterProvider` in `go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc`. (#8323)
 - `Transport` in `go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp` now supports reading request body multiple times for subsequent requests that reuse `http.Request`. (#8352)
+- Fix HTTP/1.1 connection pooling internal errors (`errTooManyIdleHost`, `errConnBroken`, `errCloseIdle`) incorrectly marking spans as errors in `go.opentelemetry.io/contrib/instrumentation/net/http/httptrace/otelhttptrace`.  (#8299)
 
 ### Changed